fix(react): harden generative UI rendering
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
